[AAP-80609] Restructure migrate_service_data tests to match module layout
Move tests from two monolithic files into module-specific test files
under tests/management/commands/test_migrate_service_data/:

- test_command.py: Command class, handle, argument parsing (10 tests)
- test_logging.py: LoggingMixin (_log, _log_progress, _configure_logging) (17 tests)
- test_legacy_authenticators.py: delete_legacy_authenticators (7 tests)
- test_resource_migration.py: resource sync, reconcile, validation (13 tests)
- test_role_assignments.py: CursorStore, pagination, assignment creation (53 tests)
- test_superuser_sync.py: superuser consistency across services (17 tests)
- test_user_merge.py: partially migrated user merging (4 tests)
- conftest.py: shared fixtures and mock helpers

Total: 121 tests (unchanged from original 77 + 44).
Update mock patch paths to target submodules where code was moved.
